+++ lib/potassium/assets/app/serializers/base_serializer.rb
@@ -1,0 +1,3 @@
+class BaseSerializer < ActiveModel::Serializer
+  include ImageHandlingAttributes
+end