+++ spec/features/file_storage_spec.rb
@@ -35,41 +35,107 @@ RSpec.describe "File Storage" do
   end
 
   context "when selecting shrine" do
-    before :all do
-      drop_dummy_database
-      remove_project_directory
-      create_dummy_project(storage: :shrine)
+    shared_examples 'common shrine setup' do
+      it "adds the aws-sdk-s3, shrine and marcel gems to Gemfile" do
+        gemfile_content = IO.read("#{project_path}/Gemfile")
+        expect(gemfile_content).to include("gem 'aws-sdk-s3'")
+        expect(gemfile_content).to include("gem 'shrine'")
+        expect(gemfile_content).to include("gem 'marcel'")
+      end
+
+      it "adds brief to README file" do
+        content = IO.read("#{project_path}/README.md")
+        expect(content).to include("Shrine")
+      end
+
+      it "adds shrine initializer" do
+        expect(File.exist?("#{project_path}/config/initializers/shrine.rb")).to be true
+      end
+
+      it "adds base_uploader and image_uploader" do
+        expect(File.exist?("#{project_path}/app/uploaders/base_uploader.rb")).to be true
+        expect(File.exist?("#{project_path}/app/uploaders/image_uploader.rb")).to be true
+      end
+
+      it "adds S3 bucket ENV vars" do
+        content = IO.read("#{project_path}/.env.development")
+        expect(content).to include("S3_BUCKET=")
+      end
+
+      it "adds filestorage path to gitignore" do
+        content = IO.read("#{project_path}/.gitignore")
+        expect(content).to include("/public/uploads")
+      end
     end
 
-    it "adds the aws-sdk-s3, shrine and marcel gems to Gemfile" do
-      gemfile_content = IO.read("#{project_path}/Gemfile")
-      expect(gemfile_content).to include("gem 'aws-sdk-s3'")
-      expect(gemfile_content).to include("gem 'shrine'")
-      expect(gemfile_content).to include("gem 'marcel'")
+    context 'with no heroku or api' do
+      before :all do
+        drop_dummy_database
+        remove_project_directory
+        create_dummy_project(storage: :shrine)
+      end
+
+      it_behaves_like 'common shrine setup'
+
+      it 'does not add buildpacks to missing .buildpacks' do
+        expect { IO.read("#{project_path}/.buildpacks") }.to raise_error(Errno::ENOENT)
+      end
+
+      it 'does not add Aptfile' do
+        expect { IO.read("#{project_path}/Aptfile") }.to raise_error(Errno::ENOENT)
+      end
+
+      it 'does not image_handling_attributes serializer concern' do
+        expect do
+          IO.read("#{project_path}/app/serializers/concerns/image_handling_attributes.rb")
+        end.to raise_error(Errno::ENOENT)
+      end
+
+      it 'does not add base serializer' do
+        expect do
+          IO.read("#{project_path}/app/serializers/base_serializer.rb")
+        end.to raise_error(Errno::ENOENT)
+      end
     end
 
-    it "adds brief to README file" do
-      content = IO.read("#{project_path}/README.md")
-      expect(content).to include("Shrine")
+    context 'with heroku' do
+      before :all do
+        drop_dummy_database
+        remove_project_directory
+        create_dummy_project(storage: :shrine, heroku: true)
+      end
+
+      it_behaves_like 'common shrine setup'
+
+      it 'adds buildpacks .buildpacks' do
+        content = IO.read("#{project_path}/.buildpacks")
+        expect(content).to include("heroku-buildpack-apt")
+        expect(content).to include("heroku-buildpack-vips")
+      end
+
+      it 'adds Aptfile' do
+        expect(File.exist?("#{project_path}/Aptfile")).to be(true)
+      end
     end
 
-    it "adds shrine initializer" do
-      expect(File.exist?("#{project_path}/config/initializers/shrine.rb")).to be true
-    end
+    context 'with rest api' do
+      before :all do
+        drop_dummy_database
+        remove_project_directory
+        create_dummy_project(storage: :shrine, api: :rest)
+      end
 
-    it "adds base_uploader and image_uploader" do
-      expect(File.exist?("#{project_path}/app/uploaders/base_uploader.rb")).to be true
-      expect(File.exist?("#{project_path}/app/uploaders/image_uploader.rb")).to be true
-    end
+      it_behaves_like 'common shrine setup'
 
-    it "adds S3 bucket ENV vars" do
-      content = IO.read("#{project_path}/.env.development")
-      expect(content).to include("S3_BUCKET=")
-    end
+      it 'adds image_handling_attributes serializer concern' do
+        expect(
+          File.exist?("#{project_path}/app/serializers/concerns/image_handling_attributes.rb")
+        ).to be(true)
+      end
 
-    it "adds filestorage path to gitignore" do
-      content = IO.read("#{project_path}/.gitignore")
-      expect(content).to include("/public/uploads")
+      it 'adds base serializer' do
+        expect(File.exist?("#{project_path}/app/serializers/base_serializer.rb")).to be(true)
+      end
     end
   end
 end