+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 Features
   - Update power api gem to use v2.0.0. Install "internal" API mode [#394](https://github.com/platanus/potassium/pull/394)
   - Updates Webpacker to Shakapacker, upgrading Vue and TailwindCSS to their latest versions [#395](https://github.com/platanus/potassium/pull/395)
+  - Add some image handling and processing in shrine file storage option [#398](https://github.com/platanus/potassium/pull/398)
 ## 6.5.0
 
 Features